+++ tests/ReplicateFilesystemAdapterTest.php
@@ -6,6 +6,7 @@ namespace Ajgl\Flysystem\Replicate\Tests;
 
 use Ajgl\Flysystem\Replicate\ReplicateFilesystemAdapter;
 use League\Flysystem\AdapterTestUtilities\FilesystemAdapterTestCase;
+use League\Flysystem\ChecksumProvider;
 use League\Flysystem\Config;
 use League\Flysystem\DirectoryAttributes;
 use League\Flysystem\FileAttributes;
@@ -175,5 +176,41 @@ final class ReplicateFilesystemAdapterTest extends FilesystemAdapterTestCase
         }
         static::$adapter = new ReplicateFilesystemAdapter(new PublicUrlGeneratorAdapter(), static::$replica);
         parent::generating_a_public_url();
+    }
+
+    /**
+     * @test
+     */
+    public function get_checksum(): void
+    {
+        if (!is_callable('parent::get_checksum')) {
+            $this->markTestSkipped();
+        }
+        static::$adapter = new ReplicateFilesystemAdapter(new ChecksumProviderAdapter(), static::$replica);
+        parent::get_checksum();
+    }
+
+    /**
+     * @test
+     */
+    public function cannot_get_checksum_for_non_existent_file(): void
+    {
+        if (!is_callable('parent::cannot_get_checksum_for_non_existent_file')) {
+            $this->markTestSkipped();
+        }
+        static::$adapter = new ReplicateFilesystemAdapter(new ChecksumProviderAdapter(), static::$replica);
+        parent::cannot_get_checksum_for_non_existent_file();
+    }
+
+    /**
+     * @test
+     */
+    public function cannot_get_checksum_for_directory(): void
+    {
+        if (!is_callable('parent::cannot_get_checksum_for_directory')) {
+            $this->markTestSkipped();
+        }
+        static::$adapter = new ReplicateFilesystemAdapter(new ChecksumProviderAdapter(), static::$replica);
+        parent::cannot_get_checksum_for_directory();
     }
 }